winter fixes for mac/linux
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2367 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Hilbert.cpp
@@ -96,6 +96,6 @@ void Hilbert::test()
 
 	for(size_t i=0; i<indices.size(); ++i)
 	{
-		conPrint("index " + ::toString(i) + ": " + toString(indices[i].first) + ", " + toString(indices[i].second));
+		conPrint("index " + ::toString((uint64)i) + ": " + toString(indices[i].first) + ", " + toString(indices[i].second));
 	}
 }
